+++ lbrynet/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-__version__ = "0.10.3production2"
+__version__ = "0.10.3production3"
 version = tuple(__version__.split('.'))
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())