+++ CHANGELOG.md
@@ -13,14 +13,12 @@ at anytime.
   *
 
 ### Changed
-  * Update `publish` to use {'currency': ..., 'address': ..., 'amount'} format for fee parameter, previously used old {currency: {'address': ..., 'amount': ...}} format
+  *
   *
 
 ### Fixed
   *
   *
-  * Allow claim_show to be used without specifying name
-  * Fix licenseUrl field in `publish`
 
 ### Deprecated
   *
@@ -29,6 +27,18 @@ at anytime.
 ### Removed
   *
   *
+
+## [0.10.3production3] - 2017-05-30
+
+### Changed
+ * Update `publish` to use {'currency': ..., 'address': ..., 'amount'} format for fee parameter, previously used old {currency: {'address': ..., 'amount': ...}} format
+
+
+### Fixed
+ * Allow claim_show to be used without specifying name
+ * Fix licenseUrl field in `publish`
+
+
 
 ## [0.10.3production2] - 2017-05-30